fix: 修复编译错误
1. 修复ServiceException为final类无法继承问题，CrmCustomerExtensionException改为继承RuntimeException
2. CrmCustomerPoolConfigDO改为继承TenantBaseDO并添加receiveLimitPerDay和receiveCooldownDays字段
3. CrmCustomerDO改为继承TenantBaseDO（支持租户隔离）
4. 修复PageResult泛型推断问题，total强制转换为Long类型

+++ Server/mitedtsm-module-crm/src/main/java/com/meession/etm/module/crm/service/customer/CrmCustomerOwnerHistoryServiceImpl.java
@@ -47,7 +47,7 @@ public class CrmCustomerOwnerHistoryServiceImpl implements CrmCustomerOwnerHisto
                 .map(this::buildRespVO)
                 .collect(Collectors.toList());
 
-        return new PageResult<>(respList, total);
+        return new PageResult<>(respList, (long) total);
     }
 
     @Override
