+++ Server/mitedtsm-module-crm/src/main/java/com/meession/etm/module/crm/service/customer/CrmHighSeasRecordServiceImpl.java
@@ -49,7 +49,7 @@ public class CrmHighSeasRecordServiceImpl implements CrmHighSeasRecordService {
                 .map(this::buildRespVO)
                 .collect(Collectors.toList());
 
-        return new PageResult<>(respList, total);
+        return new PageResult<>(respList, (long) total);
     }
 
     @Override
